[IMP] Adapt unit test
