+++ contract_variable_quantity/models/__init__.py
@@ -1,3 +1,6 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
+from . import abstract_contract_line
 from . import contract
+from . import contract_line
+from . import contract_line_formula
