+++ contract_variable_quantity/models/abstract_contract_line.py
@@ -1,0 +1,18 @@
+# Copyright 2016 Tecnativa - Pedro M. Baeza
+# Copyright 2018 Tecnativa - Carlos Dauden
+# Copyright 2018 ACSONE SA/NV
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+
+from odoo import fields, models
+
+
+class AccountAbstractAnalyticContractLine(models.AbstractModel):
+    _inherit = 'account.abstract.analytic.contract.line'
+
+    qty_type = fields.Selection(
+        selection=[
+            ('fixed', 'Fixed quantity'),
+            ('variable', 'Variable quantity'),
+        ], required=True, default='fixed', string="Qty. type")
+    qty_formula_id = fields.Many2one(
+        comodel_name="contract.line.qty.formula", string="Qty. formula")
